Added transaction to mmo_accreg_tosql, mmo_charreg_tosql, mmo_friendlist_tosql. Corrected code in mmo_party_tosql (possible ROLLBACK query without having a transaction active).
git-svn-id: http://eathena.googlecode.com/svn/branches/charmerge@13528 9c4de8d1-4af9-ca9b-867f-b47ac08f6c80

+++ src/char/accregdb_sql.c
@@ -142,19 +142,37 @@ static bool mmo_accreg_fromsql(AccRegDB_SQL* db, struct regs* reg, int account_i
 static bool mmo_accreg_tosql(AccRegDB_SQL* db, const struct regs* reg, int account_id)
 {
 	Sql* sql_handle = db->accregs;
-	SqlStmt* stmt;
-	int i;
+	SqlStmt* stmt = NULL;
+	int i = 0;
+	bool result = false;
+
+	if( SQL_SUCCESS != Sql_QueryStr(sql_handle, "START TRANSACTION") )
+	{
+		Sql_ShowDebug(sql_handle);
+		return result;
+	}
+
+	// try
+	do
+	{
 
 	//`global_reg_value` (`type`, `account_id`, `char_id`, `str`, `value`)
 	if( SQL_ERROR == Sql_Query(sql_handle, "DELETE FROM `%s` WHERE `type`=2 AND `account_id`='%d'", db->accreg_db, account_id) )
+	{
 		Sql_ShowDebug(sql_handle);
+		break;
+	}
 
 	if( reg->reg_num <= 0 )
-		return true;
+	{// nothing more needed
+		result = true;
+		break;
+	}
 
 	stmt = SqlStmt_Malloc(sql_handle);
 	if( SQL_ERROR == SqlStmt_Prepare(stmt, "INSERT INTO `%s` (`type`, `account_id`, `str`, `value`) VALUES (2,'%d',?,?)", db->accreg_db, account_id) )
 		SqlStmt_ShowDebug(stmt);
+
 	for( i = 0; i < reg->reg_num; ++i )
 	{
 		const struct global_reg* r = &reg->reg[i];
@@ -165,9 +183,29 @@ static bool mmo_accreg_tosql(AccRegDB_SQL* db, const struct regs* reg, int accou
 		SqlStmt_BindParam(stmt, 1, SQLDT_STRING, (void*)r->value, strnlen(r->value, sizeof(r->value)));
 
 		if( SQL_ERROR == SqlStmt_Execute(stmt) )
+		{
 			SqlStmt_ShowDebug(stmt);
+			break;
+		}
 	}
+
+	if( i < reg->reg_num )
+		break; // failed
+
+	// success
+	result = true;
+
+	}
+	while(0);
+	// finally
+
 	SqlStmt_Free(stmt);
 
-	return true;
+	if( SQL_SUCCESS != Sql_QueryStr(sql_handle, (result == true) ? "COMMIT" : "ROLLBACK") )
+	{
+		Sql_ShowDebug(sql_handle);
+		result = false;
+	}
+
+	return result;
 }
